Split local manifest into separate modules
